Encloses ESMF/MAPL specific GCHP interface code in ESMF pre-processor blocks.
This set of preprocessor variable changes does not affect GCHP itself,
but it allows for other model interfaces downstream (WRF & BCC) to
make use of GIGC_CHUNK_MOD in a similar way the Chem_GridComp ESMF component
does with GEOS-Chem, without having to manually merge code.

When ESMF methods or types are used alternatives are provided in the
non-ESMF variant (which may be compiled with MODEL_WRF, for example)
that have been tested to be equivalent.

Signed-off-by: Haipeng Lin <linhaipeng@pku.edu.cn>